+++ Tests/Functional/Authentication/FrontEndLoginManagerTest.php
@@ -33,6 +33,7 @@ class FrontEndLoginManagerTest extends FunctionalTestCase
     {
         parent::setUp();
         $this->testingFramework = new \Tx_Oelib_TestingFramework('tx_oelib');
+        $this->testingFramework->setResetAutoIncrementThreshold(99999999);
 
         $this->subject = \Tx_Oelib_FrontEndLoginManager::getInstance();
     }