+++ Tests/Functional/Configuration/ConfigurationCheckTest.php
@@ -38,6 +38,7 @@ class ConfigurationCheckTest extends FunctionalTestCase
         parent::setUp();
 
         $this->testingFramework = new \Tx_Oelib_TestingFramework('tx_oelib');
+        $this->testingFramework->setResetAutoIncrementThreshold(99999999);
         $this->testingFramework->createFakeFrontEnd($this->testingFramework->createFrontEndPage());
 
         $this->objectToCheck = new DummyObjectToCheck(