+++ Tests/Functional/Configuration/PageFinderTest.php
@@ -30,6 +30,7 @@ class PageFinderTest extends FunctionalTestCase
     {
         parent::setUp();
         $this->testingFramework = new \Tx_Oelib_TestingFramework('tx_oelib');
+        $this->testingFramework->setResetAutoIncrementThreshold(99999999);
 
         $this->subject = \Tx_Oelib_PageFinder::getInstance();
     }