+++ Tests/Functional/Mapper/BackEndUserMapperTest.php
@@ -31,6 +31,7 @@ class BackEndUserMapperTest extends FunctionalTestCase
     {
         parent::setUp();
         $this->testingFramework = new \Tx_Oelib_TestingFramework('tx_oelib');
+        $this->testingFramework->setResetAutoIncrementThreshold(99999999);
 
         $this->subject = new \Tx_Oelib_Mapper_BackEndUser();
     }