+++ Tests/Functional/Model/AbstractModelTest.php
@@ -44,6 +44,7 @@ class AbstractModelTest extends FunctionalTestCase
         parent::setUp();
 
         $this->testingFramework = new \Tx_Oelib_TestingFramework('tx_oelib');
+        $this->testingFramework->setResetAutoIncrementThreshold(99999999);
         \Tx_Oelib_MapperRegistry::getInstance()->activateTestingMode($this->testingFramework);
         $this->dataMapper = \Tx_Oelib_MapperRegistry::get(TestingMapper::class);